push code bill v3
