+++ project-fsoft/src/main/resources/static/statisicalstyles.css
@@ -1,0 +1,16 @@
+@charset "ISO-8859-1";
+.card-body {
+	padding: 0 20px 20px 20px;
+}
+.card-title {
+  padding: 20px 0 15px 0;
+  font-size: 18px;
+  font-weight: 500;
+  color: #012970;
+  font-family: "Poppins", sans-serif;
+}
+.card-title span {
+  color: #899bbd;
+  font-size: 14px;
+  font-weight: 400;
+}
