--- project-fsoft/src/main/resources/templates/NewFile.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="ISO-8859-1">
-<title>Insert title here</title>
-</head>
-<body>
-fffffffffffffffffff
-</body>
-</html>
